Isolate network logic from core runtime
